requirements.txt
discord.py[speed]
python-dotenv
# discord-buttons-plugin